[ENDIAN] Bit testing fixes for EventHub
This was preventing the correct flags from being set
which in turn causes mouse click events to be ignored

+++ include/ui/EventHub.h
@@ -238,7 +238,7 @@ private:
         const String8   path;
         String8         name;
         uint32_t        classes;
-        uint8_t*        keyBitmask;
+        uint32_t*       keyBitmask;
         KeyLayoutMap*   layoutMap;
         String8         keylayoutFilename;
         int             fd;
